[Feature] Event Interface
Designed around the Publisher/Subscriber philosophy.

Events exist as containers for data and nothing more.

Anything that inherits from the IEventDispatcher interface exists as an event publisher, and can dispatch events to all listeners.

Anything that inherits from the IEventListener interface exists as a subscriber, and can respond to events dispatched via a dispatcher.